pyproject.toml
[project]
name = "mips-ae-pylib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.3",
    "inquirerpy>=0.3.4",
    "pycryptodome>=3.23.0",
    "pyelftools>=0.32",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
